feat: Add summary resume on product list

+++ core/src/main/java/com/mreyesco/core/model/ProductModelInterface.kt
@@ -4,12 +4,12 @@ package com.mreyesco.core.model
 interface ProductModelInterface {
     val code: String
     val name: String
-    val price: Long
+    val price: Double
 }
 
 data class ProductModel(
     override val code: String = "",
     override val name: String = "",
-    override val price: Long = 0
+    override val price: Double = 0.0
 ) : ProductModelInterface
