+++ store/src/main/java/com/mreyesco/store/viewmodel/mapper/CurrencyMapper.kt
@@ -7,7 +7,7 @@ private const val REGEX_EMPTY: String = "\\s"
 
 object CurrencyMapper {
 
-    fun getFormattedCurrency(value: Long): String {
+    fun getFormattedCurrency(value: Double): String {
         val locale = Locale.getDefault()
         val formatter = NumberFormat.getCurrencyInstance(locale)
         return formatter.format(value).replace(REGEX_EMPTY.toRegex(), "")
